bugfix / cleardp added check for false dp

+++ modules/commands/events/dp/cleardp.js
@@ -5,6 +5,20 @@ const CONSTS = require("../../../consts");
 module.exports = {
     command: function(msg) {
         let dp = globalVariables.get("dp");
+
+        if (!dp) { // If the object is empty
+            msg.channel.send({
+                "embed": {
+                    "title": "No DP",
+                    "description": `
+                        Can't delete nothing. Or can I?
+                    `,
+                    "color": CONSTS.COLORS.RED
+                }
+            });
+            return;
+        }
+
         dp.msg.delete(); // Delete the original message
 
         globalVariables.set("dp", false);
